Refactor projects 01 through 04.
