+++ projects/02_triangle_spinning/CMakeLists.txt
@@ -1,4 +1,4 @@
-# Copyright 2022 Google LLC
+# Copyright 2023 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -17,5 +17,8 @@ project(02_triangle_spinning)
 
 add_samples_for_all_apis(
     NAME ${PROJECT_NAME}
-    SOURCES "main.cpp"
+    SOURCES
+    "TriangleSpinning.h"
+    "TriangleSpinning.cpp"
+    "main.cpp"
     SHADER_DEPENDENCIES "shader_vertex_colors")
